pyproject.toml
[project]
name = "marimo-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "geopandas>=1.1.3",
    "js>=1.0",
    "lxml>=6.0.2",
    "marimo[recommended]>=0.20.4",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "openpyxl>=3.1.5",
    "pyodide-py>=0.29.3",
    "seaborn>=0.13.2",
]
